extracted emoji hints
