+++ src/components/editor/editor-window/tool-bar/utils.ts
@@ -25,8 +25,7 @@ export const addComment = (editor: Editor): void => changeLines(editor, line => 
 export const addTable = (editor: Editor): void => changeLines(editor, line => `${line}\n| Column 1 | Column 2 | Column 3 |\n| -------- | -------- | -------- |\n| Text     | Text     | Text     |`)
 
 export const addEmoji = (emoji: EmojiData, editor: Editor): void => {
-  const replacement = getEmojiShortCode(emoji)
-  insertAtCursor(editor, replacement)
+  insertAtCursor(editor, getEmojiShortCode(emoji))
 }
 
 export const wrapTextWith = (editor: Editor, symbol: string, endSymbol?: string): void => {
